refactor(sample-next): rename "sonner" component to "toast"; refactor removeAttribute() method from analytics page
