+++ apps/sample-next/src/components/home/notifications-card.tsx
@@ -11,8 +11,8 @@ import {
 import { useOnDeviceRegistered } from "@/actito/hooks/events/core/device-registered";
 import { useOnNotificationSettingsChanged } from "@/actito/hooks/events/push/notification-settings-changed";
 import { Card, CardContent, CardHeader } from "@/components/card";
-import { toast } from "@/components/sonner";
 import { Switch } from "@/components/switch";
+import { toast } from "@/components/toast";
 import { logger } from "@/utils/logger";
 
 export function NotificationsCard() {
